create folder dedicated to post creation
